+++ NEWS.md
@@ -1,6 +1,6 @@
 # Conserve release history
 
-## v0.6.10 NOT RELEASED YET
+## v0.6.10 2020-12-30
 
 ### Features